Rename canCreateStudent middleware to isAdminOrCatchPlus
